ENG-582 hide cms menu items if not enough permissions in app-builder
